ai-worker/requirements.txt
openai==1.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-dotenv==1.0.1
tenacity==8.3.0
